refactor: 사용자 중심 설계 api controller위치 수정
