+++ src/main/java/hyundai/movie_review/review/service/ReviewService.java
@@ -219,79 +219,6 @@ public class ReviewService {
         return ReviewDetailInfoDto.of(reviewInfo, commentInfo);
     }
 
-    public ReviewInfoPageDto getReviewsByMemberId(long memberId, String type, Integer page) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(MemberIdNotFoundException::new);
-
-        Pageable pageable = PageRequest.of(page, 10);
-
-        // 1) 정렬 타입에 따라 리뷰 가져오기
-        List<Review> reviews;
-        //최신순
-        if (type.equals("latest")) {
-            reviews = reviewRepository.findByMemberAndDeletedFalseOrderByCreatedAtDesc(member,
-                    pageable);
-        }
-        //up 순
-        else if (type.equals("likes")) {
-            reviews = reviewRepository.findByMemberOrderByUps(member, pageable);
-        }
-        //별점높은순
-        else if (type.equals("ratingHigh")) {
-            reviews = reviewRepository.findByMemberAndDeletedFalseOrderByStarRateDesc(member,
-                    pageable);
-        }
-        //별점낮은순
-        else if (type.equals("ratingLow")) {
-            reviews = reviewRepository.findByMemberAndDeletedFalseOrderByStarRate(member,
-                    pageable);
-        }
-        //댓글많은순(선택)
-        else if (type.equals("comments")) {
-            reviews = reviewRepository.findByMemberOrderByComments(member, pageable);
-        } else {
-            throw new MovieReviewTypeNotFound();
-        }
-
-        // 2) 현재 로그인 한 멤버인지 확인
-        boolean isLoggedIn = memberResolver.isAuthenticated();
-        List<ReviewInfoDto> reviewInfoList;
-        if (isLoggedIn) {
-            Member currentMember = memberResolver.getCurrentMember();
-            reviewInfoList = reviews.stream()
-                    .map(review -> {
-                        boolean isThearUp = thearUpRepository.existsByMemberIdAndReviewId(
-                                currentMember,
-                                review);
-                        boolean isThearDown = thearDownRepository.existsByMemberIdAndReviewId(
-                                currentMember,
-                                review);
-                        boolean isWriter = currentMember.equals(review.getMember());
-                        return ReviewInfoDto.of(
-                                review,
-                                isThearUp,
-                                isThearDown,
-                                isWriter
-                        );
-                    }).toList();
-        } else {
-            reviewInfoList = reviews.stream()
-                    .map(review -> {
-                        return ReviewInfoDto.of(
-                                review,
-                                false,
-                                false,
-                                false
-                        );
-                    }).toList();
-        }
-
-        // 3) 페이지네이션으로 return
-        long total = reviewRepository.countByMember(member);
-        Page<ReviewInfoDto> reviewInfoPage = new PageImpl<>(reviewInfoList, pageable, total);
-        return ReviewInfoPageDto.of(reviewInfoPage);
-    }
-
     /* 영화 id에 해당하는 리뷰가 존재하는 지 검증 */
     private void validateReviewExists(Long memberId, Long movieId) {
         if (reviewRepository.existsByMemberIdAndMovieIdAndDeletedFalse(memberId, movieId)) {
